Change Guid to Int128 + Add reader/writter for the memory.
Temporary removed SmartGuid, need to reImport it properly to manage new Guid based on Int128 struct.
Note: some "wow.exe" are renamed to "wowT.exe" for the PTR version, don't forget to edit when going back to live.

+++ The Noob Bot/Quester/Tasks/QuestingTask.cs
@@ -294,7 +294,7 @@ namespace Quester.Tasks
                     if (MovementManager.InMovement)
                         return;
                     Logging.Write("Attacking Lvl " + wowUnit.Level + " " + wowUnit.Name);
-                    ulong Unkillable = Fight.StartFight(wowUnit.Guid);
+                    Int128 Unkillable = Fight.StartFight(wowUnit.Guid);
                     if (!wowUnit.IsDead && Unkillable != 0)
                     {
                         nManagerSetting.AddBlackList(Unkillable, 3*60*1000);
