+++ The Noob Bot/nManager/Wow/Bot/States/FightHostileTarget.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using nManager.FiniteStateMachine;
 using nManager.Helpful;
 using nManager.Wow.Bot.Tasks;
+using nManager.Wow.Class;
 using nManager.Wow.Enums;
 using nManager.Wow.Helpers;
 using nManager.Wow.ObjectManager;
@@ -86,7 +87,7 @@ namespace nManager.Wow.Bot.States
             if (ObjectManager.ObjectManager.Me.IsMounted)
                 return;
             Logging.Write("Player Attack " + _unit.Name + " (lvl " + _unit.Level + ")");
-            ulong unkillableMob = Fight.StartFight(_unit.Guid);
+            Int128 unkillableMob = Fight.StartFight(_unit.Guid);
             if (!_unit.IsDead && unkillableMob != 0)
             {
                 Logging.Write("Can't reach " + _unit.Name + ", blacklisting it.");
