+++ The Noob Bot/nManager/Wow/Bot/States/IsAttacked.cs
@@ -3,6 +3,7 @@ using System.Threading;
 using nManager.FiniteStateMachine;
 using nManager.Helpful;
 using nManager.Wow.Bot.Tasks;
+using nManager.Wow.Class;
 using nManager.Wow.Helpers;
 using nManager.Wow.ObjectManager;
 
@@ -73,7 +74,7 @@ namespace nManager.Wow.Bot.States
         {
             MountTask.DismountMount();
             Logging.Write("Player Attacked by " + _unit.Name + " (lvl " + _unit.Level + ")");
-            ulong unkillableMob = Fight.StartFight(_unit.Guid);
+            Int128 unkillableMob = Fight.StartFight(_unit.Guid);
             if (!_unit.IsDead && unkillableMob != 0)
             {
                 Logging.Write("Blacklisting " + _unit.Name);
